+++ packages/papa_lang/papa_lang/compiler/__init__.py
@@ -1,0 +1,23 @@
+"""papa-lang .papa DSL compiler."""
+
+from .lexer import Lexer, LexError
+from .parser import Parser, ParseError
+from .validator import Validator, ValidationError
+from .ast_nodes import AgentDef, SwarmDef, PipelineDef, Program, ConsensusConfig
+from .codegen import PythonGenerator, TypeScriptGenerator
+
+__all__ = [
+    "Lexer",
+    "LexError",
+    "Parser",
+    "ParseError",
+    "Validator",
+    "ValidationError",
+    "PythonGenerator",
+    "TypeScriptGenerator",
+    "AgentDef",
+    "SwarmDef",
+    "PipelineDef",
+    "Program",
+    "ConsensusConfig",
+]